school_management/requirements.txt
asgiref==3.8.1
colorama==0.4.6
distlib==0.3.9
Django==5.0.6
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
names==0.3.0
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary
PyJWT==2.9.0
qrcode==8.0
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.27.1
whitenoise[brotli]==6.6.0
gunicorn==21.2.0
